requirements.txt
Django==5.2.4
djangorestframework==3.16.0
django-cors-headers==4.7.0
psycopg2-binary==2.9.10
gunicorn==23.0.0
Pillow==10.0.0